requirements.txt
beautifulsoup4~=4.11.2
pwntools~=4.10.0
Requests~=2.31.0
